# .harness/contracts/openapi.yaml
openapi: 3.0.3
info:
  title: 백주대간 AI 매칭·탐색 서비스 API
  version: 0.1.0-contract
  description: >
    상태: CONFIRMED (CTR-002). 이 문서는 다음을 근거로 작성됐다:
    docs/redesign-v2/web/W-8-api-and-db-changes.md §5, docs/redesign-v2/common/
    C-5-common-api.md §1, 기존 backend/app/api/v1/endpoints/{ontology,system}.py,
    backend/app/api/health.py(BAC-001 동시 구현 중 - 이 문서는 그 계약만 문서화하고
    backend/app/api/**는 직접 수정하지 않는다), WAVE-1 §7.4 엔드포인트 목록.
    CR-001(2026-08-02) 이후 채널 전략은 WEB_ONLY로 변경 적용되어 전용 KIOSK
    채널은 GUEST_WEB 모바일 웹 세션으로 대체한다. 기존 kiosk 구현/마이그레이션은
    즉시 삭제하지 않고 deprecated 후 cleanup migration으로 처리한다.
    스키마 이름은 .harness/contracts/domain-model.md §13/§15 크로스워크를 따라 실제
    존재하는 스키마.테이블명을 반영한다(프롬프트의 평평한 테이블명이 아니라).
    오류 응답은 .harness/contracts/error-codes.yaml, 온톨로지 코드는
    .harness/contracts/ontology.yaml을 따른다.
  contact:
    name: CONTRACTS track
    url: "https://github.com/example/backju/.harness/contracts"
servers:
  - url: /api/v1
    description: 버전 프리픽스 (backend/app/core/config.py API_V1_PREFIX)
...
security:
  - bearerAuth: []
paths:
...
  /guest/sessions:
    post:
      operationId: startGuestWebSession
      summary: 게스트 웹 익명 세션 시작
      description: >
        전용 키오스크 없이 모바일 웹에서 시작하는 익명 세션. 이름·전화번호·이메일·주소,
        로그인 ID, 장기 브라우저 지문을 받지 않는다.
      tags: [guest]
      security: []
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: "#/components/schemas/StartGuestWebSessionRequest" }
      responses:
        "201":
          description: 세션 생성됨
          content:
            application/json:
              schema: { $ref: "#/components/schemas/GuestWebSessionResponse" }
  /guest/sessions/{guest_session_id}:
    get:
      operationId: getGuestWebSession
      summary: 게스트 웹 세션 상태 조회
      tags: [guest]
      security: []
      parameters:
        - name: guest_session_id
          in: path
          required: true
          schema: { type: string, format: uuid }
      responses:
        "200":
          description: 세션 상태
          content:
            application/json:
              schema: { $ref: "#/components/schemas/GuestWebSessionResponse" }
        "410":
          description: GUEST_SESSION_EXPIRED
          content:
            application/json:
              schema: { $ref: "#/components/schemas/ErrorResponse" }
...
  /guest/sessions/{guest_session_id}/favorites:
    post:
      operationId: addGuestWebFavorite
      summary: 게스트 웹 임시 관심 항목 추가
      tags: [guest, favorites]
      security: []
      parameters:
        - name: guest_session_id
          in: path
          required: true
          schema: { type: string, format: uuid }
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: "#/components/schemas/AddGuestFavoriteRequest" }
      responses:
        "201":
          description: 임시 저장됨
          content:
            application/json:
              schema: { $ref: "#/components/schemas/GuestFavoriteResponse" }
  /guest/sessions/{guest_session_id}/favorites/{temporary_favorite_id}:
    delete:
      operationId: removeGuestWebFavorite
      summary: 게스트 웹 임시 관심 항목 제거
      tags: [guest, favorites]
      security: []
      parameters:
        - name: guest_session_id
          in: path
          required: true
          schema: { type: string, format: uuid }
        - name: temporary_favorite_id
          in: path
          required: true
          schema: { type: string }
      responses:
        "204":
          description: 제거됨
  /guest/sessions/{guest_session_id}/convert:
    post:
      operationId: convertGuestWebSession
      summary: 게스트 웹 세션을 등록 사용자 프로파일에 명시적으로 연결
      description: >
        로그인/사전등록 연계 후 사용자가 명시적으로 이전을 선택한 경우에만 임시 관심목록과
        방문세션을 등록 사용자에 연결한다. 이 API 자체는 직접식별정보를 받지 않는다.
      tags: [guest, profile]
      parameters:
        - name: guest_session_id
          in: path
          required: true
          schema: { type: string, format: uuid }
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: "#/components/schemas/ConvertGuestWebSessionRequest" }
      responses:
        "200":
          description: 전환 완료
          content:
            application/json:
              schema: { $ref: "#/components/schemas/ConvertGuestWebSessionResponse" }
        "401":
          $ref: "#/components/responses/AuthRequired"
        "410":
          description: WEB_ENTRY_LINK_EXPIRED 또는 GUEST_SESSION_EXPIRED
          content:
            application/json:
              schema: { $ref: "#/components/schemas/ErrorResponse" }

  # ==========================================================================
  # EXHIBITORS/BOOTHS (C-5 §1 /common/exhibitors, /common/booths)
  # ==========================================================================
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: opaque
      description: >
        profile.user_account 인증 세션 토큰. GUEST_WEB 익명 경로는 security: []로
        명시된 엔드포인트에서만 guest_session_id(바디/경로)와 단기 토큰으로 대체 인가한다.
...
  responses:
...
    AuthRequired:
      description: AUTHENTICATION_REQUIRED
      content:
        application/json:
          schema: { $ref: "#/components/schemas/ErrorResponse" }
...
  schemas:
...
    ErrorResponse:
      type: object
      required: [error]
      properties:
        error:
          type: object
          required: [code, message, request_id]
          properties:
            code:
              type: string
              description: ".harness/contracts/error-codes.yaml http_error_codes[].code 중 하나"
              example: VALIDATION_ERROR
            message: { type: string }
            request_id: { type: string, format: uuid }
            details:
              type: object
              nullable: true
              additionalProperties: true
...
    ProfileAttributeItem:
      type: object
      properties:
        attribute_code: { type: string, description: "ontology.yaml 코드 (예: GOAL.TASTING)" }
        requirement_level:
          type: string
          enum: [REQUIRED, PREFERRED, ACCEPTABLE, EXCLUDED]
        source_type:
          type: string
          enum: [USER_SELECTED, USER_EDITED, REGISTRATION, BEHAVIOR_SINGLE, AI_EXTRACTED]
          description: "W-4 §2/§5 - 웹 모듈은 이 5종만 채운다"
    UserProfileResponse:
      type: object
      properties:
        profile_id: { type: string, format: uuid }
        user_type:
          type: string
          enum: [GENERAL_REGISTERED, BUYER_REGISTERED]
          description: "API 레이어 매핑값 - DB의 GENERAL_VISITOR/BUYER는 그대로 유지 (W-8 §2)"
        profile_status: { type: string, enum: [DRAFT, COMPLETE, INACTIVE] }
        completeness_score: { type: number, minimum: 0, maximum: 100 }
        row_version: { type: integer, description: "낙관적 동시성 - PATCH 시 If-Match로 사용" }
        attributes:
          type: array
          items: { $ref: "#/components/schemas/ProfileAttributeItem" }
        buyer_need:
          type: object
          nullable: true
          description: "user_type=BUYER_REGISTERED일 때만 존재 (profile.buyer_need)"
...
    StartGuestWebSessionRequest:
      type: object
      required: [event_id]
      properties:
        event_id: { type: string, format: uuid }
        language: { type: string, default: ko-KR }
        entry_method:
          type: string
          enum: [EVENT_WEBSITE, QR, SHORT_URL, EMAIL_LINK, REGISTRATION_LINK]
        entry_context:
          type: object
          nullable: true
          additionalProperties: true
          description: "topic QR, booth QR 등 비PII 진입 컨텍스트"
    GuestWebSessionResponse:
      type: object
      properties:
        guest_session_id: { type: string, format: uuid }
        channel: { type: string, enum: [GUEST_WEB] }
        language: { type: string }
        session_status: { type: string, enum: [ACTIVE, CONVERTED, EXPIRED, PURGED] }
        expires_at: { type: string, format: date-time }
        temporary_favorites:
          type: array
          items: { $ref: "#/components/schemas/GuestFavoriteResponse" }
    AddGuestFavoriteRequest:
      type: object
      required: [recommendable_id]
      properties:
        recommendable_id: { type: string, format: uuid }
        saved_context_json:
          type: object
          nullable: true
          additionalProperties: true
    GuestFavoriteResponse:
      type: object
      properties:
        temporary_favorite_id: { type: string }
        recommendable_id: { type: string, format: uuid }
        saved_at: { type: string, format: date-time }
    ConvertGuestWebSessionRequest:
      type: object
      properties:
        transfer_temporary_favorites: { type: boolean, default: true }
        profile_link_token:
          type: string
          nullable: true
          description: "사전등록 개인 링크/배지 QR의 단기 토큰. 원문 PII 포함 금지"
    ConvertGuestWebSessionResponse:
      type: object
      properties:
        profile: { $ref: "#/components/schemas/UserProfileResponse" }
        transferred_favorite_count: { type: integer }
        converted_at: { type: string, format: date-time }

    # ------------------------------------------------------------------------
    # EXHIBITORS/BOOTHS
    # ------------------------------------------------------------------------
...